refactor(mm): widen the loader's address plumbing to 64-bit
The multi-level walker made high user addresses mappable; every path that
would carry one was still uint32_t. try_elf_load took a uint32_t
load_base and returned uint32_t entry/img_end, the relocators took a
uint32_t slide and uint32_t seg_va[]/seg_memsz[], and the sbrk/brk
handlers computed the program break in uint32_t.

No behaviour change: nothing produces an address above 4 GiB yet, so every
value that flows through these is unchanged. That is the point of landing
it alone — the widening is mechanical and broad, the entropy change that
exercises it is neither, and a fault after both would not say which.

Two real narrowings in the ASLR source itself, both invisible at the
current window size and neither invisible after it grows:

  - aslr_rand() returned `combined >> 32`, discarding half its own output
    and capping every offset derived from it below 4 GiB. The final
    xorshift already folds the high bits down, so the low half is as mixed
    as the high; it now returns all 64 bits, with one more fold so the top
    bits are not simply dropped on the floor.
  - aslr_random_offset used `rand() % max_pages`. Modulo folds 2^64 onto
    max_pages, and unless it divides evenly the low residues get one extra
    value each — the bottom of the window is likelier than the top. At 480
    pages the skew is ~2^-55 and irrelevant; it scales with max_pages, and
    the window this is about to feed is millions of pages wide. Now
    rejection-sampled, which costs one extra draw with probability
    (2^64 mod max_pages)/2^64 and makes the distribution exact rather than
    nearly-exact-at-this-size.

The TCB's address fields were already 64-bit, so the widening holds all
the way to the page tables.

Noted, not fixed: struct task_info still carries uint32_t cr3/eip/heap_used.
It is a userspace-visible ABI (kernel, include/syscall.h and the shell must
agree byte-for-byte), and nothing truncates until images actually go high —
so it belongs with the change that moves them, not this one.

Caught by building the gated targets, not the default: ELF_SELFTEST had its
own uint32_t entry/img_end locals and the default build was perfectly
clean without them. All fourteen suites pass, plus smoke-session and a
green reproducible build.

+++ src/kernel/selftest.c
@@ -468,7 +468,7 @@ static void elf64_build_min(uint64_t e_phoff, uint64_t p_offset,
  * the test would prove nothing. */
 static int elf64_narrow_checks_ok(const char **why) {
     const uint64_t HIGH = 0x100000000ULL;   /* one bit above the 32-bit window */
-    uint32_t entry = 0, img_end = 0;
+    uint64_t entry = 0, img_end = 0;
 
     elf64_build_min(64, 0, USER_AREA_BASE, 0, 0);
     if (try_elf_load(USER_AREA_BASE, &entry, &img_end) == -17) {
@@ -734,7 +734,7 @@ void aslr_selftest(void) {
     print("ASLR_SELFTEST: begin\n");
     if (sz == 0 || sz > MAX_PROGRAM_SIZE) { print("ASLR_SELFTEST: FAIL embed-size\n"); return; }
 
-    uint32_t bases[ASLR_PROBE_SPAWNS];
+    uint64_t bases[ASLR_PROBE_SPAWNS];
     int got = 0;
     int saved = get_current_task();
